+++ CHANGELOG.md
@@ -1,7 +1,6 @@
 CHANGELOG
 ==========
 
-
 1.0.6 Bugfix release (unreleased)
 ---------------------------------
 
@@ -13,6 +12,7 @@ CHANGELOG
 #### Collectors
 
 #### Parsers
+- `bots.parsers.shadowserver`: if required fields do not exist in data, an exception is raised, so the line will be dumped and not further processed.
 
 #### Experts
 - Reverse DNS Expert: ignore all invalid results and use first valid one (#1264).